requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
locust==0.0
locustio==0.13.2
requests==2.22.0
Werkzeug==0.16.0
kafka-python==1.4.7